requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.5.0
python-dotenv==1.0.0
supabase==1.2.0
python-docx==1.1.0
rapidfuzz==3.5.2
requests==2.31.0
pydantic==2.5.2
jinja2==3.1.2
